requirements.txt
google-generativeai
python-dotenv
transformers
peft
trl
torch
datasets
wandb
opencv-python
Pillow